+++ server/routes/pages.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const { Page } = require('../models/page');
+const { auth } = require('../middleware/auth');
 
 const router = express.Router();
 
@@ -13,39 +14,55 @@ router.get('/', (req, res) => {
 router.get('/:name', (req, res) => {
   const { name } = req.params;
   Page.findOne({ name }, (err, doc) => {
-    console.log('docdocdocdocdocdocdocdocdocs', doc);
-
     if (err) return res.status(400).send(err);
     return res.send(doc);
   });
 });
 
-router.post('/', (req, res) => {
-  const page = new Page(req.body);
-  page.save((err, doc) => {
-    if (err) return res.status(400).send(err);
-    return res.status(200).json({
-      post: true,
-      pageId: doc._id,
+router.post('/', auth, (req, res) => {
+  if (req.isAuth) {
+    const page = new Page(req.body);
+    page.save((err, doc) => {
+      if (err) return res.status(400).send(err);
+      return res.status(200).json({
+        post: true,
+        pageId: doc._id,
+      });
     });
-  });
+  } else {
+    return res.status(401).json({
+      isAuth: false,
+    });
+  }
 });
 
-router.patch('/', (req, res) => {
-  Page.updateOne({ _id: req.body._id }, { $set: { ...req.body } }, (err, doc) => {
-    if (err) return res.status(400).send(err);
-    return res.status(200).json({
-      update: true,
+router.patch('/', auth, (req, res) => {
+  if (req.isAuth) {
+    Page.updateOne({ _id: req.body._id }, { $set: { ...req.body } }, (err, doc) => {
+      if (err) return res.status(400).send(err);
+      return res.status(200).json({
+        update: true,
+      });
     });
-  });
+  } else {
+    return res.status(401).json({
+      isAuth: false,
+    });
+  }
 });
 
-router.delete('/', (req, res) => {
-  const { id } = req.query;
-  Page.findByIdAndDelete(id, (err) => {
-    if (err) return res.status(400).send(err);
-    return res.json(true);
-  });
+router.delete('/', auth, (req, res) => {
+  if (req.isAuth) {
+    const { id } = req.query;
+    Page.findByIdAndDelete(id, (err) => {
+      if (err) return res.status(400).send(err);
+      return res.json(true);
+    });
+  } else {
+    return res.status(401).json({
+      isAuth: false,
+    });
+  }
 });
 
 module.exports = router;
